invoice: Update (usage) invoice code to always interpret quantity as a long (and remove int cast). See #1325
Unfortunately this is a partial fix as we would need to update our api InvoiceItem#quantity  and schema invoice_items#quantity
to the correct type as well.

+++ invoice/src/main/java/org/killbill/billing/invoice/usage/details/UsageInArrearTierUnitDetail.java
@@ -29,13 +29,13 @@ public class UsageInArrearTierUnitDetail {
     protected final int tier;
     protected final String tierUnit;
     protected final BigDecimal tierPrice;
-    protected Integer quantity;
+    protected Long quantity;
 
     @JsonCreator
     public UsageInArrearTierUnitDetail(@JsonProperty("tier") int tier,
                                        @JsonProperty("tierUnit") String tierUnit,
                                        @JsonProperty("tierPrice") BigDecimal tierPrice,
-                                       @JsonProperty("quantity") Integer quantity) {
+                                       @JsonProperty("quantity") Long quantity) {
         this.tier = tier;
         this.tierUnit = tierUnit;
         this.tierPrice = tierPrice;
@@ -54,7 +54,7 @@ public class UsageInArrearTierUnitDetail {
         return tierPrice;
     }
 
-    public Integer getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 }
